+++ src/lib.rs
@@ -8,9 +8,11 @@
 //!
 //! [`Body`]: trait.Body.html
 
+mod empty;
 mod next;
 mod size_hint;
 
+pub use self::empty::Empty;
 pub use self::next::{Data, Trailers};
 pub use self::size_hint::SizeHint;