+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+- Add `Empty`.
+
 # 0.4.0 (December 23, 2020)
 
 - Update `bytes` to v1.0.