+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+- Add combinators to `Body`:
+  - `map_data`: Change the `Data` chunks produced by the body.
+  - `map_err`: Change the `Error`s produced by the body.
+  - `boxed`: Convert the `Body` into a boxed trait object.
 - Add `Empty`.
 
 # 0.4.0 (December 23, 2020)